fix: replace router functions
