+++ src/components/cartPage.ts
@@ -3,7 +3,7 @@ import setDefaultPage from "./defaultPage";
 import getCart from "./templates/main/cart/cart";
 import {handleOrderButton} from "./functions/cart_functions/handleOrderButton";
 import {handlePromoInput} from "./functions/cart_functions/getPromoElement";
-import {getMainNavigate, goBack} from "./router/routerFuncs";
+import {getMainNavigate, goBack} from "./router/router";
 
 
 const getCartPage = () => {
