+++ src/components/router/router.ts
@@ -26,3 +26,21 @@ productData.forEach(product => {
   })
 })
 router.add('/cart', getCartPage)
+
+export function getMainNavigate() {
+  const mainPageNavigate = document.querySelector('#main');
+  if (mainPageNavigate) {
+    mainPageNavigate.addEventListener('click', () => {
+      router.navigateTo(`/`);
+    })
+  }
+}
+
+export function goBack() {
+  const back = document.querySelector('#back');
+  if (back) {
+    back.addEventListener('click', () => {
+      router.back();
+    })
+  }
+}
